+++ Changelog.md
@@ -3,6 +3,7 @@
 ## 1.0.1 - 2024.11.13
 - Modified behaviour to not crash, but skip non .NET SDK style projects
 - Only saves those projects that have been modified
+- Addef -f and --filter option to only filter out projects by name
 
 ## 1.0.0 - 2024.11.12
 - Initial release